🚚 Move: Project 관련 컴포넌트 project 패키지로 구분
