+++ src/components/project/projectGroup/ProjectGroup.style.js
