+++ src/components/title/gradientBar/GradientBar.js
