+++ src/components/title/gradientBar/GradientBar.style.js
